backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud
google-cloud-secret-manager==2.18.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# AI Integration
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
